plugins: remove usages of --experimental-type-build
