+++ .changeset/sour-toes-joke.md
@@ -1,0 +1,13 @@
+---
+'@backstage/plugin-user-settings-backend': patch
+'@backstage/plugin-scaffolder-backend': patch
+'@backstage/plugin-periskop-backend': patch
+'@backstage/plugin-scaffolder-node': patch
+'@backstage/plugin-bazaar-backend': patch
+'@backstage/plugin-kafka-backend': patch
+'@backstage/plugin-proxy-backend': patch
+'@backstage/plugin-sonarqube': patch
+'@backstage/plugin-todo': patch
+---
+
+Switched to using `"exports"` field for `/alpha` subpath export.
