+++ docs/permissions/plugin-authors/02-adding-a-basic-permission-check.md
@@ -349,7 +349,7 @@ import { createRouter } from './service/router';
 /**
 * The example TODO list backend plugin.
 *
-* @alpha
+* @public
 */
 export const exampleTodoListPlugin = createBackendPlugin({
   pluginId: 'exampleTodoList',
