+++ plugins/bazaar-backend/alpha-api-report.md
@@ -1,0 +1,13 @@
+## API Report File for "@backstage/plugin-bazaar-backend"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackendFeature } from '@backstage/backend-plugin-api';
+
+// @alpha
+const _default: () => BackendFeature;
+export default _default;
+
+// (No @packageDocumentation comment for this package)
+```
