+++ plugins/kafka-backend/api-report.md
@@ -3,17 +3,12 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { BackendFeature } from '@backstage/backend-plugin-api';
 import { Config } from '@backstage/config';
 import express from 'express';
 import { Logger } from 'winston';
 
 // @public (undocumented)
 export function createRouter(options: RouterOptions): Promise<express.Router>;
-
-// @alpha
-const kafkaPlugin: () => BackendFeature;
-export default kafkaPlugin;
 
 // @public (undocumented)
 export interface RouterOptions {
