+++ plugins/kafka-backend/src/index.ts
@@ -22,4 +22,3 @@
 
 export type { RouterOptions } from './service/router';
 export { createRouter } from './service/router';
-export { kafkaPlugin as default } from './plugin';
