+++ plugins/proxy-backend/api-report.md
@@ -3,7 +3,6 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { BackendFeature } from '@backstage/backend-plugin-api';
 import { Config } from '@backstage/config';
 import express from 'express';
 import { Logger } from 'winston';
@@ -11,10 +10,6 @@ import { PluginEndpointDiscovery } from '@backstage/backend-common';
 
 // @public
 export function createRouter(options: RouterOptions): Promise<express.Router>;
-
-// @alpha
-const proxyPlugin: () => BackendFeature;
-export default proxyPlugin;
 
 // @public (undocumented)
 export interface RouterOptions {
