+++ plugins/proxy-backend/src/index.ts
@@ -21,4 +21,3 @@
  */
 
 export * from './service';
-export { proxyPlugin as default } from './plugin';
