+++ plugins/scaffolder-node/src/index.ts
@@ -23,11 +23,3 @@
 export * from './actions';
 export * from './tasks';
 export type { TemplateFilter, TemplateGlobal } from './types';
-export {
-  scaffolderActionsExtensionPoint,
-  type ScaffolderActionsExtensionPoint,
-  scaffolderTaskBrokerExtensionPoint,
-  type ScaffolderTaskBrokerExtensionPoint,
-  scaffolderTemplatingExtensionPoint,
-  type ScaffolderTemplatingExtensionPoint,
-} from './extensions';
