+++ plugins/sonarqube/package.json
@@ -8,8 +8,7 @@
   "publishConfig": {
     "access": "public",
     "main": "dist/index.esm.js",
-    "types": "dist/index.d.ts",
-    "alphaTypes": "dist/index.alpha.d.ts"
+    "types": "dist/index.d.ts"
   },
   "backstage": {
     "role": "frontend-plugin"
@@ -27,7 +26,7 @@
   ],
   "sideEffects": false,
   "scripts": {
-    "build": "backstage-cli package build --experimental-type-build",
+    "build": "backstage-cli package build",
     "start": "backstage-cli package start",
     "lint": "backstage-cli package lint",
     "test": "backstage-cli package test",
