+++ plugins/user-settings-backend/api-report.md
@@ -3,7 +3,6 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { BackendFeature } from '@backstage/backend-plugin-api';
 import express from 'express';
 import { IdentityApi } from '@backstage/plugin-auth-node';
 import { PluginDatabaseManager } from '@backstage/backend-common';
@@ -18,10 +17,6 @@ export interface RouterOptions {
   // (undocumented)
   identity: IdentityApi;
 }
-
-// @alpha
-const userSettingsPlugin: () => BackendFeature;
-export default userSettingsPlugin;
 
 // (No @packageDocumentation comment for this package)
 ```
